Separation of concers for `retrieve_info`
The `retrieve_info` function of `info.py` now takes an argument any value that
can be used by `PIL.Image.open()`, instead of the name of the wanted station.
This forces all callers of `retrieve_info` to handle image request themselves,
and opens the door for testing by extracting local images and comparing the
output to a known transcription.